% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{add10000}
\alias{add10000}
\title{Add 10000 to the given number}
\usage{
add10000(x)
}
\arguments{
\item{x}{a number, float or int}
}
\value{
a number
}
\description{
Add 10000 to the given number
}
\examples{
add10000(5)
add10000(-90)
}
