% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_mature_mirna_reads.R
\name{collapse_mature_mirna_reads}
\alias{collapse_mature_mirna_reads}
\title{collapse_mature_mirna_reads}
\usage{
collapse_mature_mirna_reads()
}
\arguments{
\item{x}{A folder address
It can extract -3p$|-5p$|miR characteristics of non-coding RNA,
sum it's reads and it will add to the original input file
to form new file with the '_output_collapse_miRNA.txt' at the end.}
}
\value{
form new file with the '_output_collapse_miRNA.txt' at the end.
}
\description{
collapse_mature_mirna_reads
}
\examples{
collapse_mature_mirna_reads()
}
