% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_read_counts_from_sports.R
\name{combine_read_counts}
\alias{combine_read_counts}
\title{combine_read_counts_from_sports}
\usage{
combine_read_counts()
}
\arguments{
\item{x}{A folder address.}
}
\value{
form new file
sports_combined_sample_fragments_counts_matrix_all.txt
sports_combined_sample_fragments_counts_matrix_0.5.txt
sports_combined_sample_fragments_annotation.txt
}
\description{
combine_read_counts_from_sports
}
\examples{
combine_read_counts()
}
