% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdegs.R
\name{getdegs}
\alias{getdegs}
\title{DESeq2 differential analysis
get differentially expressed genes (DEGs)}
\usage{
getdegs(treat_group)
}
\arguments{
\item{treat_group}{The 11th and 12th digits (last two digits) of the name of
the SRR treatment sample group were used to define the experimental
conditions and distinguish the control group from the experimental group.}
}
\value{
Form new file
"sports_counts_all.txt","sports_DEG_fdr005_2fc_all.txt",
"sports_cpm_fdr005_2fc_all.txt","sports_DEG_all.txt"
}
\description{
DESeq2 differential analysis
get differentially expressed genes (DEGs)
}
\examples{
getdegs(20:22)
}
