% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{PCA(principal component analysis)}
\usage{
pca(treat_group)
}
\arguments{
\item{treat_group}{The 11th and 12th digits (last two digits) of the name of
the SRR treatment sample group were used to define the experimental
conditions and distinguish the control group from the experimental group.}
}
\value{
Form new file pca.pdf
}
\description{
PCA(principal component analysis)
}
\examples{
pca(20:22)
}
