% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tsRNA_mapping_plotdata_v1.2.R
\name{prepare_tsrna_mapping_plotdata}
\alias{prepare_tsrna_mapping_plotdata}
\title{prepare_tsrna_mapping_plotdata}
\usage{
prepare_tsrna_mapping_plotdata(tsRNAname, pattern = "_tRNA_mature_mapping.txt")
}
\arguments{
\item{tsRNAname}{specific tsRNA name, which you can find in "_tRNA_mature_mapping.txt".}

\item{pattern}{A uniform suffix for the filename results returned by the SPORTS1.1 tool.}
}
\value{
combined specific tsRNA mapping results, form 1 new file "specific tsRNA"_mappingplot.txt.
}
\description{
prepare_tsrna_mapping_plotdata
}
\examples{
prepare_tsrna_mapping_plotdata(tsRNAname = "Mus_musculus_mt_tRNA-Ala-TGC-1")
}
