% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_mapping_comparision.R
\name{tRNA_mapping_comparision}
\alias{tRNA_mapping_comparision}
\title{prepare_tsrna_mapping_plotdata}
\usage{
tRNA_mapping_comparision(
  filename,
  Tsamples = FALSE,
  Hsamples = FALSE,
  Lsamples = FALSE,
  Nsamples,
  pattern = "_tRNA_mature_mapping.txt"
)
}
\arguments{
\item{pattern}{A uniform suffix for the filename results returned by the SPORTS1.1 tool.}

\item{tsRNAname}{specific tsRNA name, which you can find in "_tRNA_mature_mapping.txt".}
}
\value{
combined specific tsRNA mapping results, form 1 new file "specific tsRNA"_mappingplot.txt.
}
\description{
prepare_tsrna_mapping_plotdata
}
\examples{
tRNA_mapping_comparision(filename = "Homo_sapiens_tRNA-Glu-TTC-1_mappingplot.txt",
Hsamples=c("XWH5_S7_L004", "XWH4_S6_L004", "XWH1_S0_L000", "XWH2_S6_L003"),
Lsamples=c("XWH3_S5_L004",  "XWL4_S11_L004", "XWL1_S8_L004",  "XWL3_S0_L000",  "XWL2_S9_L004",  "XWL5_S1_L001" ),
Nsamples=c("XWN1_S1_L004", "XWN2_S2_L004", "XWN3_S5_L003")
)

}
