% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsRNA_ann_classify.R
\name{tsRNA_ann_classify}
\alias{tsRNA_ann_classify}
\title{tsRNA_ann_classify}
\usage{
tsRNA_ann_classify()
}
\value{
form new file
"miRNA_diff.txt","tsRNA_diff.txt",
"rsRNA_diff.txt","ysRNA_diff.txt"
}
\description{
tsRNA_ann_classify
}
\examples{
tsRNA_ann_classify()
}
