% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_mapping_comparision.R
\name{tsRNA_mapping_comparision}
\alias{tsRNA_mapping_comparision}
\title{Plot tRNA mapping comparision coverage profile}
\usage{
tsRNA_mapping_comparision(
  filename,
  Tsamples = FALSE,
  Hsamples = FALSE,
  Lsamples = FALSE,
  Nsamples,
  pattern = "_mappingplot.txt",
  colors = c("#00AFBB", "#FC4E07", "#E7B800"),
  suffix = ".pdf"
)
}
\arguments{
\item{filename}{specific tsRNA name, which filename maybe like "_mappingplot.txt".}

\item{Tsamples}{A collection of treat group file name prefixes, generally after removing "_mappingplot.txt".}

\item{Hsamples}{A collection of high group file name prefixes, generally after removing "_mappingplot.txt".}

\item{Lsamples}{A collection of low group file name prefixes, generally after removing "_mappingplot.txt".}

\item{Nsamples}{A collection of control group file name prefixes, generally after removing "_mappingplot.txt".}

\item{pattern}{A uniform suffix for the filename results returned by the function of prepare_tsRNA_mapping_plotdata.}

\item{colors}{The color of the line in the line chart.}

\item{suffix}{A format suffix for a graph file.}
}
\value{
the coverage profile of the tsRNA and form a new file "specific tsRNA".pdf.
}
\description{
Plot the coverage profile of the tsRNA, Solid curves presenting average RPM of 2 groups(treat VS control) or 3 groups(high VS control; low VS control). RPM, reads per million; nt, nucleotide. Colored band stood for 95% CI.
}
\examples{
2 group : treat VS control
tsRNA_mapping_comparision(filename = "Homo_sapiens_tRNA-Glu-TTC-1_mappingplot.txt",
Tsamples=c("XWH5_S7_L004", "XWH4_S6_L004", "XWH1_S0_L000", "XWH2_S6_L003"),
Nsamples=c("XWN1_S1_L004", "XWN2_S2_L004", "XWN3_S5_L003")

3 group : high VS control; low VS control
tsRNA_mapping_comparision(filename = "Homo_sapiens_tRNA-Glu-TTC-1_mappingplot.txt",
Hsamples=c("XWH5_S7_L004", "XWH4_S6_L004", "XWH1_S0_L000", "XWH2_S6_L003"),
Lsamples=c("XWH3_S5_L004",  "XWL4_S11_L004", "XWL1_S8_L004",  "XWL3_S0_L000",  "XWL2_S9_L004",  "XWL5_S1_L001" ),
Nsamples=c("XWN1_S1_L004", "XWN2_S2_L004", "XWN3_S5_L003"))



}
