% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsRNA_mapping_comparision.R
\name{tsRNAmapping}
\alias{tsRNAmapping}
\title{Plot tsRNA mapping comparision coverage profile}
\usage{
tsRNAmapping(
  tsRNAname,
  pattern = "_tRNA_mature_mapping.txt",
  Tsamples = NULL,
  Hsamples = NULL,
  Lsamples = NULL,
  Nsamples,
  colors = c("#00AFBB", "#FC4E07", "#E7B800"),
  suffix = ".pdf"
)
}
\arguments{
\item{tsRNAname}{specific tsRNA name, which you can find in "_tRNA_mature_mapping.txt".}

\item{pattern}{A uniform suffix for the filename results returned by the SPORTS (Shi et al.,2018) software.}

\item{Tsamples}{A collection of treat group file name prefixes, generally after removing "_mappingplot.txt".}

\item{Hsamples}{A collection of high group file name prefixes, generally after removing "_mappingplot.txt".}

\item{Lsamples}{A collection of low group file name prefixes, generally after removing "_mappingplot.txt".}

\item{Nsamples}{A collection of control group file name prefixes, generally after removing "_mappingplot.txt".}

\item{colors}{The color of the line in the line chart.}

\item{suffix}{A format suffix for a graph file.}

\item{filename}{specific tsRNA name, which filename maybe like "_mappingplot.txt".}
}
\value{
the coverage profile of the tsRNA and form a new file "specific tsRNA".pdf.
}
\description{
Plot the coverage profile of the tsRNA, Solid curves presenting average RPM of 2 groups(treat VS control) or 3 groups(high VS control; low VS control). RPM, reads per million; nt, nucleotide. Colored band stood for 95% CI.
}
\examples{
2 group : treat VS control
tsRNAmapping(filename = "Mus_musculus_mt_tRNA-Ala-TGC-1_mappingplot.txt",
Tsamples=c("SRR11004020", "SRR11004021", "SRR11004022"),
Nsamples=c("SRR11004011", "SRR11004012", "SRR11004013"))

3 group : high VS control; low VS control
tsRNAmapping(filename = "Homo_sapiens_tRNA-Glu-TTC-1_mappingplot.txt",
Hsamples=c("XWH5_S7_L004", "XWH4_S6_L004", "XWH1_S0_L000", "XWH2_S6_L003"),
Lsamples=c("XWH3_S5_L004",  "XWL4_S11_L004", "XWL1_S8_L004",  "XWL3_S0_L000",  "XWL2_S9_L004",  "XWL5_S1_L001" ),
Nsamples=c("XWN1_S1_L004", "XWN2_S2_L004", "XWN3_S5_L003"))



}
