% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{visualization_plot}
\usage{
visualization()
}
\value{
form 5 new file pie_plot_tsRNA_aa.pdf,pie_plot_tsRNA_end.pdf,maplot.pdf,heatmap_plot.pdf,volcano_plot.pdf.
}
\description{
visualization_plot
}
\examples{
visualization()
}
