% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tsRNA_mapping_plotdata.R
\name{prepare_tsrna_mapping_plotdata}
\alias{prepare_tsrna_mapping_plotdata}
\title{Prepare tsRNA mapping plot data}
\usage{
prepare_tsrna_mapping_plotdata(tsRNAname, pattern = "_tRNA_mature_mapping.txt")
}
\arguments{
\item{tsRNAname}{specific tsRNA name, which you can find in "_tRNA_mature_mapping.txt".}

\item{pattern}{A uniform suffix for the filename results returned by the SPORTS (Shi et al.,2018) software.}
}
\value{
combined specific tsRNA mapping results, form a new file "specific tsRNA"_mappingplot.txt.
}
\description{
Prepare tsRNA mapping plot data, the data can use to plot the coverage profile of the tRNA.
}
\examples{
prepare_tsrna_mapping_plotdata(tsRNAname = "Mus_musculus_mt_tRNA-Ala-TGC-1")
}
